requirements.txt
ansible-core==2.12.5
ansible-lint==6.1.0
appdirs==1.4.4
asn1crypto==1.5.1
aspy.yaml==1.3.0
boto3==1.23.0
botocore==1.26.0
cffi==1.15.0
cfgv==3.3.1
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.2
distlib==0.3.4
docutils==0.18.1
filelock==3.7.0
identify==2.5.0
importlib-metadata==4.11.3
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
nodeenv==1.6.0
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
pep8==1.7.1
pre-commit==2.19.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
rich==12.4.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
yamllint==1.26.3
zipp==3.8.0
mitogen==0.3.0rc1
